+++ app/src/androidTest/java/me/damonkelley/tictactoe_app/adapter/BoardAdapterTest.java
@@ -1,4 +1,4 @@
-package me.damonkelley.tictactoe_app;
+package me.damonkelley.tictactoe_app.adapter;
 
 import android.support.test.filters.SmallTest;
 import android.support.test.rule.ActivityTestRule;
@@ -7,6 +7,7 @@ import android.widget.TextView;
 import me.damonkelley.tictactoe.Board;
 import me.damonkelley.tictactoe.Marker;
 import me.damonkelley.tictactoe.Space;
+import me.damonkelley.tictactoe_app.activity.GameActivity;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
