+++ app/src/main/java/me/damonkelley/tictactoe_app/loop/StateMachine.java
@@ -1,4 +1,4 @@
-package me.damonkelley.tictactoe_app;
+package me.damonkelley.tictactoe_app.loop;
 
 import me.damonkelley.tictactoe.Space;
 import me.damonkelley.tictactoe_app.turn.Turn;
