+++ app/src/main/java/me/damonkelley/tictactoe_app/wrapper/GameViews.java
@@ -1,4 +1,4 @@
-package me.damonkelley.tictactoe_app;
+package me.damonkelley.tictactoe_app.wrapper;
 
 import java.util.ArrayList;
 import java.util.List;
