+++ app/src/main/java/me/damonkelley/tictactoe_app/wrapper/ViewWrapper.java
@@ -1,4 +1,4 @@
-package me.damonkelley.tictactoe_app;
+package me.damonkelley.tictactoe_app.wrapper;
 
 public interface ViewWrapper {
     void update();
