+++ app/src/test/java/me/damonkelley/tictactoe_app/loop/LoopIntegrationTest.java
@@ -1,10 +1,12 @@
-package me.damonkelley.tictactoe_app;
+package me.damonkelley.tictactoe_app.loop;
 
 import me.damonkelley.tictactoe.Game;
 import me.damonkelley.tictactoe.Marker;
 import me.damonkelley.tictactoe.Space;
 import me.damonkelley.tictactoe.finder.ArtificialIntelligenceFinder;
 import me.damonkelley.tictactoe.finder.Finder;
+import me.damonkelley.tictactoe_app.loop.Loop;
+import me.damonkelley.tictactoe_app.loop.LoopBuilder;
 import me.damonkelley.tictactoe_app.turn.SinglePlayerHumanTurn;
 import me.damonkelley.tictactoe_app.turn.Turn;
 import org.junit.Test;
