+++ app/src/test/java/me/damonkelley/tictactoe_app/turn/SinglePlayerHumanTurnTest.java
@@ -3,7 +3,7 @@ package me.damonkelley.tictactoe_app.turn;
 import me.damonkelley.tictactoe.Game;
 import me.damonkelley.tictactoe.Marker;
 import me.damonkelley.tictactoe.Space;
-import me.damonkelley.tictactoe_app.StateMachine;
+import me.damonkelley.tictactoe_app.loop.StateMachine;
 import org.junit.Before;
 import org.junit.Test;
